Final Project Documentation first draft
We still need to make a schedule for this project, so in order to do that, please add some ideas to each object class if needed, and think about how each object can interact with others. Let me know which part of this project you'd like to work on, so we can make a schedule and assign roles and responsibilities.